Split test module
